+++ psutil/_compat.py
@@ -5,13 +5,12 @@
 """Module which provides compatibility with older Python versions."""
 
 import collections
-import contextlib
 import functools
 import os
 import sys
 
 __all__ = ["PY3", "long", "xrange", "unicode", "basestring", "u", "b",
-           "callable", "lru_cache", "which", "nested"]
+           "callable", "lru_cache", "which"]
 
 PY3 = sys.version_info[0] == 3
 
@@ -248,50 +247,3 @@ except ImportError:
                     if _access_check(name, mode):
                         return name
         return None
-
-
-# A backport of contextlib.nested for Python 3.
-nested = getattr(contextlib, "nested", None)
-if nested is None:
-    @contextlib.contextmanager
-    def nested(*managers):
-        """Support multiple context managers in a single with-statement.
-
-        Code like this:
-
-            with nested(A, B, C) as (X, Y, Z):
-                <body>
-
-        is equivalent to this:
-
-            with A as X:
-                with B as Y:
-                    with C as Z:
-                        <body>
-        """
-        exits = []
-        vars = []
-        exc = (None, None, None)
-        try:
-            for mgr in managers:
-                exit = mgr.__exit__
-                enter = mgr.__enter__
-                vars.append(enter())
-                exits.append(exit)
-            yield vars
-        except:  # NOQA
-            exc = sys.exc_info()
-        finally:
-            while exits:
-                exit = exits.pop()
-                try:
-                    if exit(*exc):
-                        exc = (None, None, None)
-                except:  # NOQA
-                    exc = sys.exc_info()
-            if exc != (None, None, None):
-                # Don't rely on sys.exc_info() still containing
-                # the right information. Another exception may
-                # have been raised and caught by an exit method
-                # exc[1] already has the __traceback__ attribute populated
-                raise exc[1]